+++ lib/tapioca/dsl/compilers/paperclip.rb
@@ -35,13 +35,15 @@ module Tapioca
         extend T::Sig
         include RBIHelper
 
-        InstanceModuleName = "PaperclipGeneratedMethods"
+        ClassMethodsModuleName = "::Paperclip::Glue"
+        InstanceMethodModuleName = "PaperclipGeneratedMethods"
+
         ConstantType = type_member { { fixed: T.class_of(::Paperclip::Glue) } }
 
         class << self
           extend T::Sig
 
-          sig { override.returns(T::Enumerable[Module]) }
+          sig { override.returns(T::Enumerable[T::Module[T.anything]]) }
           def gather_constants
             all_classes.select { |c| c < ::Paperclip::Glue }
           end
@@ -55,7 +57,7 @@ module Tapioca
           return if attachments.empty?
 
           root.create_path(constant) do |klass|
-            instance_module = RBI::Module.new(InstanceModuleName)
+            instance_module = RBI::Module.new(InstanceMethodModuleName)
 
             attachments.each do |attachment_name|
               # Model: has_attached_file(:marketing_image)
@@ -70,7 +72,8 @@ module Tapioca
             end
 
             klass << instance_module
-            klass.create_include(InstanceModuleName)
+            klass.create_include(ClassMethodsModuleName)
+            klass.create_include(InstanceMethodModuleName)
           end
         end
       end